+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `read_cmdfile_yaml` to `io` module.
+- Add implicit `--commandfile` argument to `CommandfileArgumentParser`.
 
 ## [0.1.3] - 2025-08-07